+++ RELEASE_NOTES.md
@@ -380,6 +380,8 @@ Akka.NET v1.5.48 is a minor patch containing stability improvement to Akka.TestK
 
 To [see the full set of changes in Akka.NET v1.5.48, click here](https://github.com/akkadotnet/akka.net/milestone/131?closed=1)
 
+* Core: Add `ILoggingAdapter` context enrichment, explicit scopes, and bracketed context output in StandardOutLogger and Xunit logger
+
 #### 1.5.47 August 12th, 2025 ####
 
 Akka.NET v1.5.47 is a minor patch containing several stability improvements to Akka.TestKit.